fixup: move `isOnboardingUser` to dashboard
